+++ core-auth-fe/src/core-auth/resource/identities/ConnectSocialLogin.tsx
@@ -2,17 +2,7 @@ import { Box, Input } from '@mui/material'
 import { useState } from 'react'
 import { SocialSignInProps, standard } from '../../components/login-mui/SocialSignIn'
 import { defaultCredentials } from '../../credentials'
-import { StyledIcon } from '@styled-icons/styled-icon'
 import { csrfToken } from '../../react-admin/LoginPage'
-import { MoreIcons } from '../../components/MoreIcons'
-
-const { Apple, Facebook, Google, Microsoft } = MoreIcons
-const identityProviders: { [k: string]: StyledIcon } = {
-  google: Google,
-  apple: Apple,
-  facebook: Facebook,
-  microsoft: Microsoft,
-}
 
 type OauthProvider = {
   k: string
@@ -26,7 +16,7 @@ const social = defaultCredentials.getAppConfig()?.oauth_providers?.enabled.map((
 
 const csrf = csrfToken()
 
-const socialSignInRender: SocialSignInProps['renderButton'] = ({ key, icon, label, href, disabled }) => {
+const socialSignInRender: SocialSignInProps['renderButton'] = ({ icon, label, href, disabled }) => {
   return (
     <Box
       component='form'
@@ -40,10 +30,12 @@ const socialSignInRender: SocialSignInProps['renderButton'] = ({ key, icon, labe
 }
 
 import { Button, Menu, MenuItem, Fade } from '@mui/material'
+import { RaRecord, useListContext } from 'react-admin'
 
-export const ConnectSocialLogin = (props: any) => {
+export const ConnectSocialLogin = () => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
   const open = Boolean(anchorEl)
+  const { data } = useListContext()
   
   const handleClick = (event: React.MouseEvent<HTMLElement>) => {
     setAnchorEl(event.currentTarget)
@@ -53,7 +45,7 @@ export const ConnectSocialLogin = (props: any) => {
     setAnchorEl(null)
   }
 
-  if (!props.data) return <></>
+  if (!data) return null
 
   return (
     <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
@@ -77,7 +69,7 @@ export const ConnectSocialLogin = (props: any) => {
         TransitionComponent={Fade}
       >
         {social.map(({ k, href }: { k: any, href: any }) => {
-          const disabled = !!props.data.find((record: any) => record.provider === k)
+          const disabled = !!data.find((record: RaRecord) => record.provider === k)
 
           const s = standard.find((s) => s.key == k)
           if (!s) return null
@@ -85,8 +77,8 @@ export const ConnectSocialLogin = (props: any) => {
           const { key, icon, label } = s
 
           return (
-            <MenuItem>
-              <span style={{ width: '100%' }} key={key}>{socialSignInRender({ key, icon, label, href: `${href}?return_to=${window.location.href}`, disabled })}</ span>
+            <MenuItem key={key}>
+              <span style={{ width: '100%' }}>{socialSignInRender({ key, icon, label, href: `${href}?return_to=${window.location.href}`, disabled })}</ span>
             </MenuItem>    
           )
         })}